Made close dialog look similar to that of the other modals. Also refactored SurveyWrapper
