+++ jhlongcovid/components/Header/ProfileCheckin.tsx
@@ -73,7 +73,14 @@ export const Survey: React.FC<ProfileCheckinProps> = ({
       >
         Participate
       </Button>
-      <Modal isOpen={isOpen} onClose={closeSurvey} isCentered size={"lg"} closeOnEsc={false} closeOnOverlayClick={false}>
+      <Modal
+        isOpen={isOpen}
+        onClose={closeSurvey}
+        isCentered
+        size={"lg"}
+        closeOnEsc={false}
+        closeOnOverlayClick={false}
+      >
         <ModalOverlay p="30px" />
         <SurveyWrapper onClose={closeSurvey} />
         {/* {showSurvey && <SurveyWrapper onClose={onClose} />} */}
