+++ jhlongcovid/components/Survey/SurveyBody/MFA.tsx
@@ -15,7 +15,7 @@ import {
   FormHelperText,
   FormErrorMessage,
 } from "@chakra-ui/react";
-import { SurveyQuestionProps } from "../SurveyWrapper";
+import { SurveyQuestionProps } from "../SurveyBody";
 
 export const MFA: React.FC<SurveyQuestionProps> = ({
   currentQuestion,
