+++ jhlongcovid/tsconfig.json
@@ -13,8 +13,7 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "jsx": "preserve",
-    "incremental": true,
-    "types": ["jest", "node", "@types/testing-library__jest-dom"]
+    "incremental": true
   },
   "include": [
     "next-env.d.ts",
